update destroy function

+++ app/Http/Controllers/Admin/PositionController.php
@@ -3,11 +3,15 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\ResponseTrait;
 use App\Models\Position;
+use App\Models\User;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Str;
 class PositionController extends Controller
 {
+    use ResponseTrait;
+
     public function index()
     {
         $positions = Position::query()
@@ -25,16 +29,22 @@ class PositionController extends Controller
 
     public function store(Request $request)
     {
-        $check = Position::query()  
+        $count_pos = Position::selectRaw('count(id) as count')->value('count');
+        if($count_pos == 5)
+        {
+            return redirect()->back()->with('error', 'Tối đa chỉ được 4 chức vụ!!!');
+        }
+        $pos_name = Str::title($request->position_name);
+        $check_null = Position::query()  
                 ->where('name', $request->position_name)
                 ->first();
-        if(!is_null($check))
+        if(!is_null($check_null))
         {
-            return redirect()->back()->with('error', 'Mon da ton tai trong he thong');
+            return redirect()->back()->with('error', 'Chức vụ đã tồn tại trong hệ thống');
         }
 
         Position::create([
-            'name' => $request->position_name,
+            'name' => $pos_name,
             'salary' => $request->price * 1000,
         ]);
 
@@ -66,7 +76,27 @@ class PositionController extends Controller
 
     public function destroy(Request $request)
     {
+        $pos_id = $request->pos_id;
+
+        // $check = User::query()
+        //         ->where('role', $pos_id)
+        //         ->value('name');
+
+        // $check = User::query()
+        //         ->where('role', $pos_id)
+        //         ->get();
+        $check = User::query()
+                ->where('role', $pos_id)
+                ->exists();
+        // dd($check);
+        if($check == true)
+        {
+            return $this->errorResponse('Hãy đảm bảo không còn nhân viên nào thuộc chức vụ này! Hãy thử lại sau');
+        }
         Position::destroy($request->pos_id);
-        return redirect()->back();
+        // return redirect()->back();
+        return $this->successResponse([
+            'pos_id' => $pos_id,
+        ], 'Xoá chức vụ thành công!');
     }
 }
